# lambdas/tickets-aws-lambda/internal/adapters/handlers/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ticketera Cloud Tickets Microservice API",
        "title": "Tickets API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3006",
    "basePath": "/",
    "paths": {
        "/tickets/me": {
            "get": {
                "description": "Retrieve real-time purchased tickets for the logged user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get user tickets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Ticket"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Ticket": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "event_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                },
                "status": {
                    "description": "PAID, PROCESSING",
                    "type": "string"
                }
            }
        }
    }
}